Bullets 이미지 및 이펙트 현질 및 적용ㅋ

+++ Assets/Scripts/UI/SubItem/UI_TextEquipStatusValue.cs
@@ -20,7 +20,8 @@ public class UI_TextEquipStatusValue : UI_Base
         }
         else
         {
-            valueText += ((value -1) * 100) + "%";
+            valueText += (int)((value - 1) * 100) + "%";
+            
         }
         GetComponent<TextMeshProUGUI>().text = valueText;
     }
